requirements.txt
Flask==3.0.3
beautifulsoup4==4.12.3
requests==2.32.3
websocket-client==1.8.0
